+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Versioned according to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+  * Adapt to feature selection/filtering mechanism for derived images in core
+  * Fixes for image feature-related corner cases in crop and deskew
+  * Use explicit (second) output fileGrp when producing derived images
+
 ## [0.4.0] - 2019-08-21
 
 Changed: